+++ packages/node/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Subcommands for `force-clean` and `reindex`. (#1281)
+- `yargs` file has been moved back into `node` from `node-core`. (#1281)
 
 ## [1.9.2] - 2022-09-19
 ### Changed